image_engine/requirements.txt
numpy>=1.24
Pillow>=10.0
scipy>=1.10
imageio>=2.30
